# perf(ts): McapStreamDispatcher, a faster iterator